+++ src/main/kotlin/kr/kro/dearmoment/inquiry/adapter/input/web/ArtistInquiryRestAdapter.kt
@@ -1,10 +1,10 @@
-package kr.kro.dearmoment.inquiry.adapter.input.web.artist
+package kr.kro.dearmoment.inquiry.adapter.input.web
 
 import kr.kro.dearmoment.common.dto.PagedResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.artist.dto.CreateArtistInquiryRequest
-import kr.kro.dearmoment.inquiry.adapter.input.web.artist.dto.GetArtistInquiryResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.dto.CreateInquiryResponse
 import kr.kro.dearmoment.inquiry.application.command.CreateArtistInquiryCommand
+import kr.kro.dearmoment.inquiry.application.dto.CreateArtistInquiryRequest
+import kr.kro.dearmoment.inquiry.application.dto.CreateInquiryResponse
+import kr.kro.dearmoment.inquiry.application.dto.GetArtistInquiryResponse
 import kr.kro.dearmoment.inquiry.application.port.input.CreateInquiryUseCase
 import kr.kro.dearmoment.inquiry.application.port.input.GetInquiryUseCase
 import kr.kro.dearmoment.inquiry.application.query.GetArtistInquiresQuery
