+++ src/main/kotlin/kr/kro/dearmoment/inquiry/adapter/input/web/ProductInquiryRestAdapter.kt
@@ -1,10 +1,10 @@
-package kr.kro.dearmoment.inquiry.adapter.input.web.product
+package kr.kro.dearmoment.inquiry.adapter.input.web
 
 import kr.kro.dearmoment.common.dto.PagedResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.dto.CreateInquiryResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.product.dto.CreateProductInquiryRequest
-import kr.kro.dearmoment.inquiry.adapter.input.web.product.dto.GetProductInquiryResponse
 import kr.kro.dearmoment.inquiry.application.command.CreateProductInquiryCommand
+import kr.kro.dearmoment.inquiry.application.dto.CreateInquiryResponse
+import kr.kro.dearmoment.inquiry.application.dto.CreateProductInquiryRequest
+import kr.kro.dearmoment.inquiry.application.dto.GetProductInquiryResponse
 import kr.kro.dearmoment.inquiry.application.port.input.CreateInquiryUseCase
 import kr.kro.dearmoment.inquiry.application.port.input.GetInquiryUseCase
 import kr.kro.dearmoment.inquiry.application.port.input.RemoveInquiryUseCase
