+++ src/main/kotlin/kr/kro/dearmoment/inquiry/application/dto/CreateArtistInquiryRequest.kt
@@ -1,19 +1,18 @@
-package kr.kro.dearmoment.inquiry.adapter.input.web.service.dto
+package kr.kro.dearmoment.inquiry.application.dto
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotBlank
 import org.jetbrains.annotations.NotNull
 
-data class CreateServiceInquiryRequest(
+data class CreateArtistInquiryRequest(
     @field:NotNull(value = "유저 ID 널이 될 수 없습니다.")
     val userId: Long,
-    @field:NotBlank(message = "서비스 피드백 타입은 빈 문자열이 될 수 없습니다.")
-    val type: String,
+    @field:NotBlank(message = "제목은 빈 문자열이 될 수 없습니다.")
+    @field:Max(value = 15L, message = "제목은 최대 15자 이하여야합니다.")
+    val title: String,
     @field:NotBlank(message = "내용은 빈 문자열이 될 수 없습니다.")
-    @field:Min(value = 10, message = "내용은 최소 10자 이상이어야 합니다.")
-    @field:Max(value = 1000, message = "내용은 최소 1000자 이하이어야 합니다.")
+    @field:Max(value = 200L, message = "내용은 최대 200자 이하여야합니다.")
     val content: String,
     @field:Email(message = "올바른 이메일 형식이 아닙니다.")
     val email: String,
