+++ src/main/kotlin/kr/kro/dearmoment/inquiry/application/dto/CreateInquiryResponse.kt
@@ -1,4 +1,4 @@
-package kr.kro.dearmoment.inquiry.adapter.input.web.dto
+package kr.kro.dearmoment.inquiry.application.dto
 
 data class CreateInquiryResponse(
     val inquiryId: Long,
