+++ src/main/kotlin/kr/kro/dearmoment/inquiry/application/dto/CreateProductInquiryRequest.kt
@@ -1,4 +1,4 @@
-package kr.kro.dearmoment.inquiry.adapter.input.web.product.dto
+package kr.kro.dearmoment.inquiry.application.dto
 
 import org.jetbrains.annotations.NotNull
