+++ src/main/kotlin/kr/kro/dearmoment/inquiry/application/dto/GetProductInquiryResponse.kt
@@ -1,4 +1,4 @@
-package kr.kro.dearmoment.inquiry.adapter.input.web.product.dto
+package kr.kro.dearmoment.inquiry.application.dto
 
 import kr.kro.dearmoment.inquiry.domain.ProductInquiry
 import java.time.LocalDateTime
