+++ src/main/kotlin/kr/kro/dearmoment/inquiry/application/port/input/CreateInquiryUseCase.kt
@@ -1,9 +1,9 @@
 package kr.kro.dearmoment.inquiry.application.port.input
 
-import kr.kro.dearmoment.inquiry.adapter.input.web.dto.CreateInquiryResponse
 import kr.kro.dearmoment.inquiry.application.command.CreateArtistInquiryCommand
 import kr.kro.dearmoment.inquiry.application.command.CreateProductInquiryCommand
 import kr.kro.dearmoment.inquiry.application.command.CreateServiceInquiryCommand
+import kr.kro.dearmoment.inquiry.application.dto.CreateInquiryResponse
 
 interface CreateInquiryUseCase {
     fun createArtistInquiry(command: CreateArtistInquiryCommand): CreateInquiryResponse
