+++ src/main/kotlin/kr/kro/dearmoment/inquiry/application/port/input/GetInquiryUseCase.kt
@@ -1,8 +1,8 @@
 package kr.kro.dearmoment.inquiry.application.port.input
 
 import kr.kro.dearmoment.common.dto.PagedResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.artist.dto.GetArtistInquiryResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.product.dto.GetProductInquiryResponse
+import kr.kro.dearmoment.inquiry.application.dto.GetArtistInquiryResponse
+import kr.kro.dearmoment.inquiry.application.dto.GetProductInquiryResponse
 import kr.kro.dearmoment.inquiry.application.query.GetArtistInquiresQuery
 import kr.kro.dearmoment.inquiry.application.query.GetProductInquiresQuery
