+++ src/test/kotlin/kr/kro/dearmoment/inquiry/adapter/input/web/product/ProductInquiryRestAdapterTest.kt
@@ -19,9 +19,9 @@ import kr.kro.dearmoment.common.restdocs.requestBody
 import kr.kro.dearmoment.common.restdocs.responseBody
 import kr.kro.dearmoment.common.restdocs.toJsonString
 import kr.kro.dearmoment.common.restdocs.type
-import kr.kro.dearmoment.inquiry.adapter.input.web.dto.CreateInquiryResponse
-import kr.kro.dearmoment.inquiry.adapter.input.web.product.dto.CreateProductInquiryRequest
-import kr.kro.dearmoment.inquiry.adapter.input.web.product.dto.GetProductInquiryResponse
+import kr.kro.dearmoment.inquiry.application.dto.CreateInquiryResponse
+import kr.kro.dearmoment.inquiry.application.dto.CreateProductInquiryRequest
+import kr.kro.dearmoment.inquiry.application.dto.GetProductInquiryResponse
 import kr.kro.dearmoment.inquiry.application.query.GetProductInquiresQuery
 import org.junit.jupiter.api.Test
 import org.springframework.data.domain.PageImpl
